Split Configuration components up
